Release version 1.0.1
The minimum supported version with all features is 2020.1. In older Unity versions, the editor and tests assemblies will not be compiled (silently ignored). Also, if the test package is not installed, the comment component tests are also ignored. And finally, some small warnings in older versions of Unity are avoided by assigning default values to serialized members.